разбивает файл на модули
